refactor regex pattern for stop regex
