+++ rust/sglang-server/src/lib.rs
@@ -21,6 +21,7 @@ mod ring;
 mod runtime;
 mod tokenizer;
 mod tokenizer_manager;
+mod utils;
 
 use std::net::SocketAddr;
