+++ rust/sglang-server/src/utils.rs
@@ -1,0 +1,3 @@
+//! Shared helpers with no home in a pipeline stage.
+
+pub mod regex;
